functions/requirements.txt
google-cloud-firestore
google-auth
google-auth-oauthlib
google-api-python-client
firebase-functions
firebase-admin
